requirements.txt
flask
flask-cors
edge-tts
pygame
pyttsx3
pywin32
pyautogui
screen-brightness-control
comtypes
pycaw
pystray
Pillow
numpy
windows-capture
